services/brain-core/requirements.txt
fastapi~=0.115.0
uvicorn[standard]~=0.32.0
pydantic>=2.11.0
pydantic-settings>=2.7.0
httpx~=0.28.0
anthropic~=0.42.0
redis~=5.2.0
asyncpg~=0.30.0
pymilvus~=2.4.0
elasticsearch~=8.15.0
langchain~=0.3.0
langchain-community~=0.3.0
prometheus-client~=0.21.0
python-dotenv~=1.0.0
numpy>=1.26.0
edge-tts~=7.0.0
python-multipart~=0.0.20
Pillow>=10.0.0
qrcode~=7.4
pydub~=0.25.0
slowapi~=0.1.9
structlog~=24.4.0
openpyxl>=3.1.0
python-docx>=1.1.0
PyMuPDF>=1.24.0
mcp>=1.26.0
jieba>=0.42.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
